% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regcut.R
\name{.regcut}
\alias{.regcut}
\alias{lm.regcut}
\alias{coeftest.regcut}
\title{Get Results of Regression}
\usage{
lm.regcut(
  reg,
  keep = NULL,
  omit = NULL,
  digits = NULL,
  intercept.include = NULL,
  star = NULL
)

coeftest.regcut(
  reg,
  keep = NULL,
  omit = NULL,
  digits = NULL,
  intercept.include = NULL,
  star = c("*", "**", "***")
)
}
\arguments{
\item{reg}{regression with specific class}

\item{keep}{keep variable name}

\item{omit}{omit variable name}

\item{digits}{maximum digits}

\item{intercept.include}{logical value including intercept}

\item{star}{vector of symbols representing statistically significance at 10%, 5%, and 1% level}
}
\value{
generate column of coefficient of regression
}
\description{
Get Results of Regression
}
