% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_flextable.R
\name{reg.flextable}
\alias{reg.flextable}
\title{Generate Regression Table}
\usage{
reg.flextable(
  ...,
  robust = NULL,
  title = NULL,
  label = NULL,
  dep.var.caption = NULL,
  dep.var.caption.separate = NULL,
  column.labels = NULL,
  column.separate = NULL,
  keep = NULL,
  omit = NULL,
  intercept.include = TRUE,
  order = NULL,
  covariate.labels = NULL,
  add.lines = NULL,
  df = TRUE,
  digits = 2,
  notes = NULL,
  font.size = 12,
  autofit = TRUE,
  covariate.width = 1.5,
  reg.width = 1,
  minor.change,
  star = c("*", "**", "***"),
  keep.stat = c("n")
)
}
\arguments{
\item{...}{list of regressions which extract regression stats}

\item{robust}{list of regressions which extract regression coefficients}

\item{title}{character of table title}

\item{label}{character of table label which returns (#tab:"label"). When you use the cross-reference, you should use this option}

\item{dep.var.caption}{a character vector of captions for dependent variables in regression tables.}

\item{dep.var.caption.separate}{a numeric vector that specifies how dep.var.caption should be laid out across regression table columns.}

\item{column.labels}{a character vector of labels for columns in regression tables. Their layout, in terms of the number of columns associated with each label, is given by the argument column.separate.}

\item{column.separate}{a numeric vector that specifies how column.labels should be laid out across regression table columns. A value of c(2, 1, 3), for instance, will apply the first label to the two first columns, the second label to the third column, and the third label will apply to the following three columns (i.e., columns number four, five and six).}

\item{keep}{a vector of regular expressions that specifies which of the explanatory variables should be kept in the table.}

\item{omit}{a vector of regular expressions that specifies which of the explanatory variables should be omitted in the table.}

\item{intercept.include}{logical value including intercept}

\item{order}{a vector of numerical indexes that indicates the order in which variables will appear in the output.}

\item{covariate.labels}{a character vector of labels for covariates in regression tables. A value of NA for any element means that stargazer will print the corresponding variable name.}

\item{add.lines}{a list containing vectors that shows additional lines.}

\item{df}{a logical value that indicates whether the degrees of freedom of model statistics should be reported.}

\item{digits}{an integer that indicates how many decimal places should be used.}

\item{notes}{a character vector containing notes to be included below the table.}

\item{font.size}{an integer that specifies the font size used in the table.}

\item{autofit}{a logical value that indicates whether autofit function in flextable works (Strongly recommend TRUE)}

\item{covariate.width}{a numeric value of width of covariate column (inches)}

\item{reg.width}{a numeric value of width of regression column (inches)}

\item{minor.change}{a function that change specific style of table}

\item{star}{a character vector symbols representing statistically significance at 10\%, 5\%, and 1\% level}

\item{keep.stat}{a character vector that specifies which model statistics should be kept in the regression table output. For instance keep.stat = c("n","adj.rsq") will produce a table that only includes statistics for the number of observations and Adjusted R-squared.}
}
\value{
object "flextable" with regression table
}
\description{
Generate Regression Table
}
