% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regcuttable.R
\name{regcuttable}
\alias{regcuttable}
\title{Generate table of regression coefficients}
\usage{
regcuttable(regcut, covariate.labels = NULL, order = NULL)
}
\arguments{
\item{regcut}{list of regression coef data generated by .regcut function}

\item{covariate.labels}{a character vector of labels for covariates in regression tables. A value of NA for any element means that stargazer will print the corresponding variable name.}

\item{order}{a vector of numerical indexes that indicates the order in which variables will appear in the output.}
}
\value{
data.frame which contains regression coefficients
}
\description{
Generate table of regression coefficients
}
