% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reginfotable.R
\name{reginfotable}
\alias{reginfotable}
\title{Make table of regression statistics}
\usage{
reginfotable(reginfo)
}
\arguments{
\item{reginfo}{list of regression statistics extracted by .reginfo function}
}
\value{
data.frame of regression statistics
}
\description{
Make table of regression statistics
}
